Refactor into a history recording system
